pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask-dance>=7.1.0",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "flask-login>=0.6.3",
    "oauthlib>=3.3.1",
    "pillow>=11.3.0",
    "sqlalchemy>=2.0.43",
    "werkzeug>=3.1.3",
    "requests>=2.32.5",
    "replit-object-storage>=1.0.2",
]
